chore/PAS-2 rewrite create order
