+++ apps/frontend/src/app/CreateOrder/index.ts
@@ -1,1 +1,1 @@
-export * from './CreateOrder';
+export * from './create-order.component';
